moved operation provider to impl, cleaned leftovers, changed back to NodeEngine from NodeEngineImpl
