+++ hazelcast/src/main/java/com/hazelcast/cache/impl/DefaultOperationProvider.java
@@ -1,6 +1,5 @@
 package com.hazelcast.cache.impl;
 
-import com.hazelcast.cache.CacheOperationProvider;
 import com.hazelcast.cache.impl.operation.CacheClearOperationFactory;
 import com.hazelcast.cache.impl.operation.CacheContainsKeyOperation;
 import com.hazelcast.cache.impl.operation.CacheEntryProcessorOperation;
