+++ hazelcast/src/main/java/com/hazelcast/cache/impl/HazelcastServerCachingProvider.java
@@ -32,6 +32,7 @@ import java.util.Properties;
  * created managers.</p>
  * <p>If you need to use your already created HazelcastInstance, you can directly create a provider using
  * {@link #createCachingProvider(com.hazelcast.core.HazelcastInstance)}.</p>
+ *
  * @see javax.cache.spi.CachingProvider
  */
 public final class HazelcastServerCachingProvider
@@ -54,7 +55,6 @@ public final class HazelcastServerCachingProvider
         if (config.getInstanceName() == null) {
             config.setInstanceName("CacheProvider");
         }
-        config.getNetworkConfig().getJoin().getMulticastConfig().setEnabled(false);
         return HazelcastInstanceFactory.getOrCreateHazelcastInstance(config);
     }
